End of Session 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();    
@@ -31,15 +32,19 @@ export class MessagesComponent implements OnInit {
     })
   }
 
+  deleteMessage(id: number) {
+    this.confirmService.confirm('Confirm Delete', 'Delete Messages cannot be undo, do you want to continue?').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(i => i.id === id), 1);
+        })
+      }
+    });   
+  }
+
   pageChanged(event: any) {
     this.pageNumber = event.page;
     this.loadMessages();
-  }
-
-  deleteMessage(id: number)   {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(i => i.id === id), 1);
-    })
   }
 
 }
